Fix issue with Coverity inside Docker by workaround
Attempts to fix the broken wrapper are futile. The correct fix is to use
a proper Workflow Step and not a SimpleBuildWrapper as this is doing.
The changes would be very invasive and not worth my time.

Instead we assume the responsibility of using a properly installed
Coverity instance inside the configured Docker image and a compatible
version of Coverity server. These are the onyl features we give up when
disabling the Validate step in the workflow configuration here.

The changes are small enough that hopefully can continue moving forward
with not much effort on our part until Coverity Synopsys decides to
take action on this the right way.